Add files via upload
Car Rental System made by Daryl James is a Desktop-based program that uses Java Swing and MySQL, designed to make and manage reservations in car rentals easier. This is a system that streamlines vehicle management, payment handling made from reservation, admin management (exclusive to the owner), and verifying admin. This system is solely made for educational purposes and is not intended for real-life renting.